ai-chaos-engineer/browser-use-service/requirements.txt
browser-use>=0.9.6
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0

